[ETHEREUM-CONTRACTS/SDK-CORE] CFAv1 Forwarder Integration (#1118)
* Agreement Forwarder Integration WIP

* Split supertoken tests into separate files
* Add CFAv1Forwarder to SuperfluidFrameworkDeployer + set in resolver and enable as trusted forwarder
* Integrate CFAv1Forwarder in SDK-Core
* tests to ensure that forwarder is being used and works as expected
* validateOperationShouldUseCallAgreement checks that we are using agreement forwarder

* Update CHANGELOG.md

* SDK-Core cleanup


* Fix them hardcoded addresses!

* two step fix

* Subgraph test setup cleanup

* delete test-subgraph.template, it is a duplicate
* rename prepare-local: prepare-manifest-local
* update ganache.json
* create runDeployContractsAndToken.ts to deploy framework and tokens and create new ganache.json dynamically
* no more hardcoded addresses in addresses.template.ts!
* helpers.ts modified to read ganache.json file to retrieve resolver address to setup framework

* must build sdk-core

* Fixes

* Make it properly dynamic by deploying and preparing local files prior to actual subgraph deployment
* fix up package.json given the above
* fix up readme

* build sdk-core before deploy contracts

* do not redeploy

* missed the other reusable workflow

* chainId unddefined?

* sender is not needed here
