+++ .github/workflows/call.test-subgraph-on-previous-sdk-core-versions.yml
@@ -44,6 +44,19 @@ jobs:
         run: |
           ./tasks/startHardhatNode.sh start
         working-directory: ${{ env.sdk-core-working-directory }}
+      
+      - name: "Build SDK-Core"
+        # build sdk-core because of auto linking to dependency
+        run: yarn build
+        working-directory: ${{ env.sdk-core-working-directory }}
+
+      - name: "Deploy Framework and Tokens"
+        run: npx hardhat run scripts/runDeployContractsAndToken.ts --network localhost
+        working-directory: ./packages/subgraph
+
+      - name: "Prepare files for local testing"
+        run: yarn prepare-local
+        working-directory: ./packages/subgraph
 
       # duplicated in call.setup-deploy-and-test-local-subgraph
       - name: "Checkout graph node repo and set up local graph node"
