+++ .github/workflows/handler.deploy-subgraph.yml
@@ -44,7 +44,7 @@ jobs:
         working-directory: ${{ env.subgraph-working-directory }}
 
       - name: "Generate subgraph manifest"
-        run: "yarn prepare-local" # NOTE: A hack to make codegen work
+        run: "yarn prepare-manifest-local" # NOTE: A hack to make codegen work
         working-directory: ${{ env.subgraph-working-directory }}
 
       - name: "Generate meta.ignore.ts file"
