+++ packages/subgraph/config/ganache.json
@@ -1,10 +1,1 @@
-{
-    "network": "mainnet",
-    "testNetwork": "ganache",
-    "hostStartBlock": 0,
-    "hostAddress": "0xE451980132E65465d0a498c53f0b5227326Dd73F",
-    "cfaAddress": "0x5392A33F7F677f59e833FEBF4016cDDD88fF9E67",
-    "idaAddress": "0xa783CDc72e34a174CCa57a6d9a74904d0Bec05A9",
-    "superTokenFactoryAddress": "0x61ef99673A65BeE0512b8d1eB1aA656866D24296",
-    "resolverV1Address": "0xF45bcaDCc83dea176213Ae4E22f5aF918d08647b"
-}
+{"network":"mainnet","testNetwork":"ganache","hostStartBlock":0,"hostAddress":"0xE451980132E65465d0a498c53f0b5227326Dd73F","cfaAddress":"0x5392A33F7F677f59e833FEBF4016cDDD88fF9E67","idaAddress":"0xB30dAf0240261Be564Cea33260F01213c47AAa0D","superTokenFactoryAddress":"0xF45bcaDCc83dea176213Ae4E22f5aF918d08647b","resolverV1Address":"0xeCaE6Cc78251a4F3B8d70c9BD4De1B3742338489","nativeAssetSuperTokenAddress":"0x51f9613a79D56528622Bd31a5Fd4b88b78AE4F8A"}
