+++ packages/subgraph/package.json
@@ -21,12 +21,17 @@
         "integrity": "npx hardhat run scripts/dataIntegrity/dataIntegrityTest.ts --network",
         "check-indexing-completeness": "ts-node scripts/checkIsDeployedOnAllNetworks.ts",
         "deploy-contracts-local": "npx hardhat run scripts/deployContractsAndToken.ts --network localhost",
-        "prepare-local": "mustache config/ganache.json test-subgraph.template.yaml > subgraph.yaml",
+        "prepare-manifest-local": "mustache config/ganache.json subgraph.template.yaml > subgraph.yaml",
         "set-network-local": "mustache config/ganache.json src/addresses.template.ts > src/addresses.ts",
         "create-local": "graph create superfluid-test --node http://localhost:8020/",
         "remove-local": "graph remove superfluid-test --node http://localhost:8020/",
         "deploy-local": "graph deploy superfluid-test --node http://localhost:8020/ --ipfs http://localhost:5001 --version-label v1.0.0",
-        "build-and-deploy-local": "yarn prepare-local && yarn set-network-local && yarn getAbi && yarn generate-sf-meta-local && yarn codegen && yarn create-local && yarn deploy-local",
+        "prepare-local": "run-s prepare-local:*",
+        "prepare-local:manifest": "yarn prepare-manifest-local",
+        "prepare-local:network": "yarn set-network-local",
+        "prepare-local:abi": "yarn getAbi",
+        "prepare-local:sf-meta": "yarn generate-sf-meta-local",
+        "build-and-deploy-local": "yarn codegen && yarn create-local && yarn deploy-local",
         "watch": "graph deploy superfluid-test --node http://localhost:8020/ --ipfs http://localhost:5001 --watch",
         "deploy": "chmod +x ./tasks/deploy.sh && ./tasks/deploy.sh",
         "deploy-subgraph": "graph deploy $SUBGRAPH_NAME --node https://api.thegraph.com/deploy/ --ipfs https://api.thegraph.com/ipfs --access-token $THEGRAPH_ACCESS_TOKEN",
